Fix crashes when quitting the game
They happen when there's global idDicts that are destroyed at exit.
Like that static idDict nameAliases; (which I replaced with a simple
array of structs) and idDict members of gameLocal (which is also a
global variable so its destructor is called on exit).
The fix for gameLocal is to call Clear() on all its dicts in
idGameLocal.Clear() (which is called by idGameLocal::Shutdown()).

The crashes happen because uncleared idDicts reference strings from
idStrPool which gets cleared by idLib::Shutdown() -> idDict::Shutdown()