Fix crash in idRadioCheckin::Think()
soundCue is a const char*, idStr::Format() returns an idStr temporary,
which is invalid (out of scope) in the very next line. So you either
need to assign it to another idStr (or some similar kind that actually
copies the string), or you need to use something else.

va() has internal (static) char buffers that it writes to and returns,
and it cycles through 16 of them so it will take some times until the
one returned in one call gets reused by another va() call, so it should
be safe here (soundCue is only used within this function)