Hack to get rid of GL_ARB_shading_language_420pack dependency
The hacky part of this is modifying the shader source in readFile().
It works with the current shaders, but different formatting would
break that hack.

Note that this change does *NOT* make this compatible with macOS,
because SkinDeep still requires an OpenGL >= 3.3 *Compatibility* Profile
but macOS only supports OpenGL Core Profiles (or OpenGL 2 legacy mode,
which doesn't work with SkinDeep either).
But this might help with older GPUs and their drivers.

Leaving it in a separate branch because it's hacky and so far no one
needed it (as it doesn't help with macOS).