Make OpenOSFile less spammy if developer 1
It'll usually log several "can't open" messages for every single file,
because the game first tries to load it from the config dir, then from
the save dir, then from the install dir..
and in case of .wav files it tries .ogg first and in case of .tga it
tries .dds first so that's a whole lotta files that can't be opened when
loading a map.
Now this message is hidden behind `fs_debug 2`.