Set "noclipmodel" spawn arg of particles to true
If this isn't set, it will try to load a collisionmodel later in
idEntity::InitDefaultPhysics() (and fail, because it doesn't exist).
This is done every time a particle is spawned that way, which also
happens frequently during runtime, so this saves quite some pointless
work (including trying to open particleName.cm at several different
locations).

This was especially noticeable when setting `developer 1` in
vig_surveil.map which then spammed the console with messages like
"idFileSystem::OpenOSFile: could not open file due to err 2:
 /home/alexander/.config/skindeep/base/spaceparticles64_forcezerog.cm"